+++ example/midway-react-ssr/src/controller/index.ts
@@ -1,10 +1,9 @@
-import { Readable } from 'stream'
 import { Controller, Get, Provide, Inject } from '@midwayjs/decorator'
-import { Context } from 'egg'
+import { Context } from '@midwayjs/koa'
 import { render } from 'ssr-core-react'
 import { IApiService, IApiDetailService } from '../interface'
 
-interface IEggContext extends Context {
+interface IKoaContext extends Context {
   apiService: IApiService
   apiDeatilservice: IApiDetailService
 }
@@ -13,7 +12,7 @@ interface IEggContext extends Context {
 @Controller('/')
 export class Index {
   @Inject()
-  ctx: IEggContext
+  ctx: IKoaContext
 
   @Inject('ApiService')
   apiService: IApiService
@@ -27,7 +26,7 @@ export class Index {
     try {
       this.ctx.apiService = this.apiService
       this.ctx.apiDeatilservice = this.apiDeatilservice
-      const stream = await render<Readable>(this.ctx, {
+      const stream = await render(this.ctx, {
         stream: true
       })
       this.ctx.body = stream