+++ packages/client-utils/src/preload.ts
@@ -1,17 +1,15 @@
-// @ts-nocheck
-import { ReactClientESMFeRouteItem } from 'ssr-types'
+import { ReactESMPreloadFeRouteItem, ReactESMFeRouteItem } from 'ssr-types-react'
+import { normalizePath } from 'ssr-common-utils'
 import { pathToRegexp } from 'path-to-regexp'
-import { normalizePath } from './utils'
-
-const preloadComponent = async (Routes: ReactClientESMFeRouteItem[], PrefixRouterBase?: string) => {
+const preloadComponent = async (Routes: ReactESMPreloadFeRouteItem[], base?: string) => {
   for (const route of Routes) {
     const { component, path } = route
     let pathname = location.pathname
-    if (PrefixRouterBase) {
-      pathname = normalizePath(pathname, PrefixRouterBase)
+    if (base) {
+      pathname = normalizePath(pathname, base)
     }
     if (component.name === 'dynamicComponent' && pathToRegexp(path).test(pathname)) {
-      route.component = (await component()).default
+      route.component = (await (component as ReactESMFeRouteItem['component'])()).default
     }
   }
   return Routes