+++ packages/plugin-vue/src/entry/client-entry.ts
@@ -1,12 +1,11 @@
 import { Store } from 'vuex'
 import { Route } from 'vue-router'
-import { findRoute } from 'ssr-client-utils'
+import { findRoute, isMicro } from 'ssr-client-utils'
 import { Routes } from './create-router'
 import { ESMFetch, RoutesType, IFeRouteItem } from './interface'
 import { createRouter, createStore, RealVue } from './create'
 
-declare const module: any
-const { FeRoutes, App, layoutFetch, PrefixRouterBase } = Routes as RoutesType
+const { FeRoutes, App, layoutFetch } = Routes as RoutesType
 
 let hasRender = false
 async function getAsyncCombineData (fetch: ESMFetch | undefined, store: Store<any>, router: Route) {
@@ -25,23 +24,29 @@ async function getAsyncCombineData (fetch: ESMFetch | undefined, store: Store<an
 const clientRender = async () => {
   const store = createStore()
   const router = createRouter({
-    base: window.prefix ?? PrefixRouterBase
+    base: isMicro() ? window.clientPrefix : window.prefix
   })
 
   if (window.__INITIAL_DATA__) {
     store.replaceState(window.__INITIAL_DATA__)
   }
   const fetchData = window.__INITIAL_DATA__ ?? {}
-
+  const reactiveFetchData = {
+    value: window.__INITIAL_DATA__ ?? {}
+  }
   const app = new RealVue({
     // 根实例简单的渲染应用程序组件。
     render: h => h(App, {
       props: {
-        fetchData
+        fetchData,
+        reactiveFetchData
       }
     }),
     store,
-    router
+    // for type checker
+    ...{
+      router
+    }
   })
 
   router.beforeResolve(async (to, from, next) => {
@@ -55,6 +60,7 @@ const clientRender = async () => {
           fetchData: combineAysncData
         })
       })
+      reactiveFetchData.value = combineAysncData
     }
     hasRender = true
     next()
@@ -63,17 +69,9 @@ const clientRender = async () => {
   router.onReady(() => {
     app.$mount('#app', !!window.__USE_SSR__) // 这里需要做判断 ssr/csr 来为 true/false
   })
-
-  if (!window.__USE_VITE__) {
-    module?.hot?.accept?.() // webpack 场景下的 hmr
-  }
 }
 
-if (!window.__disableClientRender__) {
-  // 如果服务端直出的时候带上该记号，则默认不进行客户端渲染，将处理逻辑交给上层
-  // 可用于微前端场景下自定义什么时候进行组件渲染的逻辑调用
-  clientRender()
-}
+clientRender()
 
 export {
   clientRender