+++ packages/server-utils/src/loadConfig.ts
@@ -1,9 +1,11 @@
 import { join } from 'path'
 import { IConfig } from 'ssr-types'
-import { getCwd, getUserConfig, normalizeStartPath, normalizeEndPath, getFeDir, judgeFramework, loadModuleFromFramework, stringifyDefine, accessFileSync } from './cwd'
+import { normalizeStartPath, normalizeEndPath } from 'ssr-common-utils'
+import { getCwd, getUserConfig, getFeDir, judgeFramework, loadModuleFromFramework, stringifyDefine, accessFileSync } from './cwd'
 import { coerce } from 'semver'
-const framework = judgeFramework()
+
 const loadConfig = (): IConfig => {
+  const framework = judgeFramework()
   const userConfig = getUserConfig()
   const cwd = getCwd()
   const mode = 'ssr'
@@ -33,8 +35,8 @@ const loadConfig = (): IConfig => {
   }, userConfig.alias)
 
   type ClientLogLevel = 'error'
-
   const publicPath = userConfig.publicPath?.startsWith('http') ? userConfig.publicPath : normalizeStartPath(userConfig.publicPath ?? '/')
+
   const devPublicPath = publicPath.startsWith('http') ? publicPath.replace(/^http(s)?:\/\/(.*)?\d/, '') : publicPath // 本地开发不使用 http://localhost:3000 这样的 path 赋值给 webpack-dev-server 会很难处理
 
   const moduleFileExtensions = [
@@ -52,10 +54,13 @@ const loadConfig = (): IConfig => {
     '.vue',
     '.css'
   ]
-
   const isDev = userConfig.isDev ?? process.env.NODE_ENV !== 'production'
 
-  const fePort = userConfig.fePort ?? 8888
+  const fePort = userConfig.fePort ?? 8999
+
+  const hmr = Object.assign({
+    host: '127.0.0.1'
+  }, userConfig.hmr)
 
   let https = userConfig.https ? userConfig.https : !!process.env.HTTPS
 
@@ -65,19 +70,19 @@ const loadConfig = (): IConfig => {
 
   const serverPort = process.env.SERVER_PORT ? Number(process.env.SERVER_PORT) : 3000
 
-  const host = '0.0.0.0'
+  const host = hmr?.host ?? '127.0.0.1'
 
   const chunkName = 'Page'
 
   const clientLogLevel: ClientLogLevel = 'error'
 
   const useHash = !isDev // 生产环境默认生成hash
+  const defaultWhiteList: Array<RegExp|string> = [/\.(css|less|sass|scss)$/, /vant.*?style/, /antd.*?(style)/, /ant-design-vue.*?(style)/, /store$/, /\.(vue)$/]
+  const whiteList: Array<RegExp|string> = defaultWhiteList.concat(userConfig.whiteList ?? [])
 
-  const whiteList: RegExp[] = [/\.(css|less|sass|scss)$/, /vant.*?style/, /antd.*?(style)/, /ant-design-vue.*?(style)/, /store$/]
+  const jsOrder = isVite ? [`${chunkName}.js`] : [`runtime~${chunkName}.js`, 'vendor.js', `${chunkName}.js`]
 
-  const jsOrder = isVite ? [`${chunkName}.js`] : [`runtime~${chunkName}.js`, 'vendor.js', `${chunkName}.js`].concat(userConfig.extraJsOrder ?? [])
-
-  const cssOrder = ['vendor.css', `${chunkName}.css`].concat(userConfig.extraCssOrder ?? [])
+  const cssOrder = ['vendor.css', `${chunkName}.css`]
 
   const webpackStatsOption = {
     assets: true, // 添加资源信息
@@ -122,9 +127,10 @@ const loadConfig = (): IConfig => {
     publicPath: devPublicPath,
     hotOnly: true,
     host,
-    sockPort: fePort,
+    sockHost: host,
+    sockPort: hmr?.port ?? fePort,
     hot: true,
-    port: fePort,
+    port: hmr?.port ?? fePort,
     https,
     clientLogLevel: clientLogLevel,
     headers: {
@@ -148,7 +154,7 @@ const loadConfig = (): IConfig => {
   const staticPath = `${normalizeEndPath(devPublicPath)}static`
   const hotUpdatePath = `${normalizeEndPath(devPublicPath)}*.hot-update**`
   const proxyKey = [staticPath, hotUpdatePath, manifestPath]
-
+  const prefix = '/'
   const config = Object.assign({}, {
     chainBaseConfig,
     chainServerConfig,
@@ -166,7 +172,6 @@ const loadConfig = (): IConfig => {
     cssOrder,
     getOutput,
     webpackStatsOption,
-    whiteList,
     dynamic,
     mode,
     stream,
@@ -180,15 +185,18 @@ const loadConfig = (): IConfig => {
     reactServerEntry,
     reactClientEntry,
     isVite,
+    whiteList,
     isCI,
     supportOptinalChaining,
-    define
+    define,
+    prefix
   }, userConfig)
   config.alias = alias
+  config.prefix = normalizeStartPath(config.prefix ?? '/')
   config.corejsOptions = corejsOptions
-
+  config.whiteList = whiteList
+  config.hmr = hmr
   config.webpackDevServerConfig = webpackDevServerConfig // 防止把整个 webpackDevServerConfig 全量覆盖了
-
   config.babelOptions = userConfig.babelOptions ? {
     ...{
       babelHelpers: 'bundled' as 'bundled',
@@ -197,6 +205,7 @@ const loadConfig = (): IConfig => {
     },
     ...userConfig.babelOptions
   } : undefined
+
   return config
 }